refactor(renderer): extract launcher command execution helpers
